fix: move heatmap fetch into component and revert stat maxValue to 100
